display error

+++ mod/arkeogis/Main.php
@@ -221,6 +221,18 @@ class Main {
 				$geographical_limit_de = trim($form->getValue('geographical_limit_de'));
 				if (!empty($geographical_limit_de))
 					$field['geographical_limit_de'] = $geographical_limit_de;
+				if ($file['error']) {
+					$errs=array(1=>"The uploaded file exceeds the upload_max_filesize directive in php.ini",
+											2=>"The uploaded file exceeds the MAX_FILE_SIZE directive that was specified in the HTML form",
+											3=>"The uploaded file was only partially uploaded",
+											4=>"No file was uploaded",
+											6=>"Missing a temporary folder.",
+											7=>"Failed to write file to disk",
+											8=>"A PHP extension stopped the file upload");
+					$err=isset($errs[$file['error']]) ? $errs[$file['error']] : $file['error'];
+					self::display_html_error(\mod\lang\Main::ch_t('arkeogis', "Une erreur est survenue lors de l'envoi du fichier : ").$err);
+					return;
+				}
 				$result =	\mod\arkeogis\DatabaseImport::importCsv($file['tmp_name'], $separator, $enclosure, $skipline, $lang, $field);
 				unlink($file['tmp_name']);
 				$page = new \mod\webpage\Main();
